Factor out the action functions.
Factor out the non-UI action functions.
